Refactor for MVC. Changed package name from "panel" to "view"
