+++ src/main/java/tetris/view/GameLoop.java
@@ -1,4 +1,4 @@
-package tetris.panel;
+package tetris.view;
 
 import javafx.animation.AnimationTimer;
